LPS-121592 Move to new package to avoid excessive semver bumps
