+++ portal-kernel/src/com/liferay/portal/kernel/resource/bundle/CacheResourceBundleLoader.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
+package com.liferay.portal.kernel.resource.bundle;
 
 import com.liferay.petra.concurrent.ConcurrentReferenceKeyHashMap;
 import com.liferay.petra.memory.FinalizeManager;
@@ -29,11 +29,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * @author Carlos Sierra Andrés
- *
- * @deprecated As of Athanasius (7.3.x), replaced by {@link
- *             com.liferay.portal.kernel.resource.bundle.CacheResourceBundleLoader}
  */
-@Deprecated
 public class CacheResourceBundleLoader implements ResourceBundleLoader {
 
 	public static void clearCache() {
