+++ portal-kernel/src/com/liferay/portal/kernel/resource/bundle/ResourceBundleLoader.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
+package com.liferay.portal.kernel.resource.bundle;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
@@ -21,11 +21,7 @@ import org.osgi.annotation.versioning.ProviderType;
 
 /**
  * @author Carlos Sierra Andrés
- *
- * @deprecated As of Athanasius (7.3.x), replaced by {@link
- *             com.liferay.portal.kernel.resource.bundle.ResourceBundleLoader}
  */
-@Deprecated
 @FunctionalInterface
 @ProviderType
 public interface ResourceBundleLoader {
