+++ portal-kernel/src/com/liferay/portal/kernel/resource/bundle/ResourceBundleLoaderUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
+package com.liferay.portal.kernel.resource.bundle;
 
 import com.liferay.registry.ServiceReference;
 import com.liferay.registry.collections.ServiceReferenceMapper;
@@ -21,11 +21,7 @@ import com.liferay.registry.collections.ServiceTrackerMap;
 
 /**
  * @author Carlos Sierra Andrés
- *
- * @deprecated As of Athanasius (7.3.x), replaced by {@link
- *             com.liferay.portal.kernel.resource.bundle.ResourceBundleLoaderUtil}
  */
-@Deprecated
 public class ResourceBundleLoaderUtil {
 
 	public static ResourceBundleLoader getPortalResourceBundleLoader() {
